DjangoApi/Trending/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.4.0
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
lxml==4.5.2
psycopg2==2.7.7
psycopg2-binary==2.8.6
pytz==2020.1
requests==2.24.0
selenium==3.141.0
soupsieve==2.0.1
sqlparse==0.3.1
tqdm==4.49.0
urllib3==1.25.10
webdriver-manager==3.2.2
webdrivermanager==0.9.0
whitenoise==5.2.0
